TwitterSA/src/main/java/com/ceccoemi/twittersa/Trainer.java
package com.ceccoemi.twittersa;

import com.aliasi.classify.Classification;
import com.aliasi.classify.Classified;
import com.aliasi.classify.DynamicLMClassifier;
import com.aliasi.lm.NGramProcessLM;

import java.io.FileOutputStream;
import java.io.IOException;
import java.io.ObjectOutputStream;
import java.util.Iterator;

public class Trainer {

  private DynamicLMClassifier<NGramProcessLM> classifier =
      DynamicLMClassifier.createNGramProcess(new String[]{"1", "0"}, 8);

  public void train(Iterator<Tweet> tweets) {
    while (tweets.hasNext()) {
      Tweet tweet = tweets.next();
      String text = tweet.getText();
      String sentiment = tweet.getSentiment();
      Classification classification = new Classification(sentiment);
      Classified<CharSequence> classified = new Classified<CharSequence>(text, classification);
      classifier.handle(classified);
    }
  }

  public void storeModel(String fileName) throws IOException {
      FileOutputStream fos = new FileOutputStream(fileName);
      ObjectOutputStream oos = new ObjectOutputStream(fos);
      classifier.compileTo(oos);
  }

}
